+++ Realm/ObjectServerTests/SwiftObjectServerTests.swift
@@ -21,18 +21,6 @@ import RealmSwift
 
 class SwiftObjectServerTests: SwiftSyncTestCase {
 
-    func workAroundCoreIssue2724() {
-        // As described in https://github.com/realm/realm-core/issues/2724, with realm-core v3.0.0-rc3 we're
-        // seeing occasional deadlocks in these tests due to core thinking that the history schema needs to
-        // be upgraded when it doesn't. This leads to it taking a write transaction as part of opening a file.
-        // Since this can happen on the notification listener thread while another thread both has an existing
-        // write transaction and is waiting for the notification listener thread to process notifications,
-        // we deadlock.
-        // We work around this by delaying slightly between registering for notifications on a given Realm and
-        // opening a write transaction on the same Realm.
-        RunLoop.current.run(until: Date(timeIntervalSinceNow: 0.1))
-    }
-
     /// It should be possible to successfully open a Realm configured for sync.
     func testBasicSwiftSync() {
         let url = URL(string: "realm://localhost:9080/~/testBasicSync")!
@@ -99,7 +87,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
         }
     }
 
-    // MARK: Client reset
+    // MARK: - Client reset
 
     func testClientReset() {
         do {
@@ -159,7 +147,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
         }
     }
 
-    // MARK: Progress notifiers
+    // MARK: - Progress notifiers
 
     func testStreamingDownloadNotifier() {
         let bigObjectCount = 2
@@ -242,7 +230,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
         }
     }
 
-    // MARK: Download Realm
+    // MARK: - Download Realm
 
     func testDownloadRealm() {
         let bigObjectCount = 2
@@ -292,7 +280,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
         }
     }
 
-    // MARK: Administration
+    // MARK: - Administration
 
     func testRetrieveUserInfo() {
         let nonAdminUsername = "meela.swift@realm.example.org"
@@ -322,7 +310,63 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
         waitForExpectations(timeout: 10.0, handler: nil)
     }
 
-    // MARK: Permissions
+    // MARK: - Authentication
+
+    func testInvalidCredentials() {
+        do {
+            let username = "testInvalidCredentialsUsername"
+            let credentials = SyncCredentials.usernamePassword(username: username,
+                                                               password: "THIS_IS_A_PASSWORD",
+                                                               register: true)
+            _ = try synchronouslyLogInUser(for: credentials, server: authURL)
+            // Now log in the same user, but with a bad password.
+            let ex = expectation(description: "wait for user login")
+            let credentials2 = SyncCredentials.usernamePassword(username: username, password: "NOT_A_VALID_PASSWORD")
+            SyncUser.logIn(with: credentials2, server: authURL) { user, error in
+                XCTAssertNil(user)
+                XCTAssertTrue(error is SyncAuthError)
+                let castError = error as! SyncAuthError
+                XCTAssertEqual(castError.code, SyncAuthError.invalidCredential)
+                ex.fulfill()
+            }
+            waitForExpectations(timeout: 2.0, handler: nil)
+        } catch {
+            XCTFail("Got an error: \(error) (process: \(isParent ? "parent" : "child"))")
+        }
+    }
+
+    // MARK: - User-specific functionality
+
+    func testUserExpirationCallback() {
+        do {
+            let user = try synchronouslyLogInUser(for: basicCredentials(), server: authURL)
+
+            // Set a callback on the user
+            let ex = expectation(description: "Error callback should fire upon receiving an error")
+            var invoked = false
+            user.errorHandler = { (u, error) in
+                XCTAssertEqual(u.identity, user.identity)
+                XCTAssertEqual(error.code, .invalidCredential)
+                invoked = true
+                ex.fulfill()
+            }
+
+            // Screw up the token on the user.
+            manuallySetRefreshToken(for: user, value: "not-a-real-token")
+
+            // Try to open a Realm with the user; this will cause our errorHandler block defined above to be fired.
+            _ = try immediatelyOpenRealm(url: realmURL, user: user)
+            if !invoked {
+                waitForExpectations(timeout: 10.0, handler: nil)
+            }
+            XCTAssertEqual(user.state, .loggedOut)
+
+        } catch {
+            XCTFail("Got an error: \(error) (process: \(isParent ? "parent" : "child"))")
+        }
+    }
+
+    // MARK: - Permissions
 
     func testPermissionOffer() {
         do {