+++ Realm/RLMNetworkClient.mm
@@ -19,7 +19,7 @@
 #import "RLMNetworkClient.h"
 
 #import "RLMRealmConfiguration.h"
-#import "RLMSyncErrorResponseModel.h"
+#import "RLMJSONModels.h"
 #import "RLMSyncUtil_Private.hpp"
 #import "RLMUtil.hpp"
 
@@ -256,6 +256,11 @@ static NSRange RLM_rangeForErrorType(RLMServerHTTPErrorCodeType type) {
                 case RLMSyncAuthErrorInvalidCredential:
                 case RLMSyncAuthErrorUserDoesNotExist:
                 case RLMSyncAuthErrorUserAlreadyExists:
+                case RLMSyncAuthErrorAccessDeniedOrInvalidPath:
+                case RLMSyncAuthErrorInvalidAccessToken:
+                case RLMSyncAuthErrorExpiredPermissionOffer:
+                case RLMSyncAuthErrorAmbiguousPermissionOffer:
+                case RLMSyncAuthErrorFileCannotBeShared:
                     *error = make_auth_error(responseModel);
                     break;
                 default: