+++ Realm/Tests/RLMTestCase.h
@@ -33,13 +33,6 @@ NSData *RLMGenerateKey(void);
 }
 #endif
 
-#if __has_attribute(swift_objc_members)
-#define REALM_SWIFT_OBJC_MEMBERS __attribute__((swift_objc_members))
-#else
-#define REALM_SWIFT_OBJC_MEMBERS
-#endif
-
-REALM_SWIFT_OBJC_MEMBERS // workaround for rdar://33514802
 @interface RLMTestCase : XCTestCase
 
 - (RLMRealm *)realmWithTestPath;